websocket import issue.
"ws" module is commonJS supporting.
it has an "export = WebSocket", which is not properly handled by nodejs ESModule imports.
we need to enable the typescript esModuleInterop property in the tsconfg.json file for that.
for this we need to initialize a ts thingy in the nodejs backend.
and then compile the ws library and then include it from the dist/ folder produced after compilation.
now everything is js.
node runtime can run this js code now :)